Move tests into chain specific folders
