+++ packages/app/src/features/dialogs/savings/deposit/e2e/mainnet/DepositDAI.test-e2e.ts
@@ -1,18 +1,14 @@
 import { ActionsPageObject } from '@/features/actions/ActionsContainer.PageObject'
 import { SavingsPageObject } from '@/pages/Savings.PageObject'
-import { PSM_ACTIONS_DEPLOYED, PSM_ACTIONS_DEPLOYED_DATE } from '@/test/e2e/constants'
+import { DEFAULT_BLOCK_NUMBER } from '@/test/e2e/constants'
 import { setup } from '@/test/e2e/setup'
 import { setupFork } from '@/test/e2e/setupFork'
 import { test } from '@playwright/test'
 import { mainnet } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../common/e2e/SavingsDialog.PageObject'
+import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
 
-test.describe('Deposit USDC on Mainnet', () => {
-  const fork = setupFork({
-    blockNumber: PSM_ACTIONS_DEPLOYED,
-    simulationDateOverride: PSM_ACTIONS_DEPLOYED_DATE,
-    chainId: mainnet.id,
-  })
+test.describe('Deposit DAI on Mainnet', () => {
+  const fork = setupFork({ blockNumber: DEFAULT_BLOCK_NUMBER, chainId: mainnet.id })
   let savingsPage: SavingsPageObject
   let depositDialog: SavingsDialogPageObject
 
@@ -23,47 +19,43 @@ test.describe('Deposit USDC on Mainnet', () => {
         type: 'connected',
         assetBalances: {
           ETH: 1,
-          USDC: 10_000,
+          DAI: 10_000,
         },
       },
     })
 
     savingsPage = new SavingsPageObject(page)
-    await savingsPage.clickDepositButtonAction('USDC')
+    await savingsPage.clickDepositButtonAction('DAI')
 
     depositDialog = new SavingsDialogPageObject({ page, type: 'deposit' })
     await depositDialog.fillAmountAction(10_000)
   })
 
-  test('uses PSM actions native deposit', async () => {
+  test('uses native sDai deposit', async () => {
     await depositDialog.actionsContainer.expectActions([
-      { type: 'approve', asset: 'USDC' },
-      { type: 'usdcToSDaiDeposit', asset: 'USDC' },
+      { type: 'approve', asset: 'DAI' },
+      { type: 'daiToSDaiDeposit', asset: 'DAI' },
     ])
   })
 
   test('displays transaction overview', async () => {
     await depositDialog.expectNativeRouteTransactionOverview({
       apy: {
-        value: '8.00%',
-        description: '~800.00 DAI per year',
+        value: '5.00%',
+        description: '~500.00 DAI per year',
       },
       routeItems: [
-        {
-          tokenAmount: '10,000.00 USDC',
-          tokenUsdValue: '$10,000.00',
-        },
         {
           tokenAmount: '10,000.00 DAI',
           tokenUsdValue: '$10,000.00',
         },
         {
-          tokenAmount: '9,196.30 sDAI',
+          tokenAmount: '9,332.66 sDAI',
           tokenUsdValue: '$10,000.00',
         },
       ],
-      outcome: '9,196.30 sDAI worth $10,000.00',
-      badgeToken: 'USDC',
+      outcome: '9,332.66 sDAI worth $10,000.00',
+      badgeToken: 'DAI',
     })
   })
 
@@ -74,7 +66,7 @@ test.describe('Deposit USDC on Mainnet', () => {
     await depositDialog.expectSuccessPage()
     await depositDialog.clickBackToSavingsButton()
 
-    await savingsPage.expectSavingsBalance({ sDaiBalance: '9,196.30 sDAI', estimatedDaiValue: '10,000' })
-    await savingsPage.expectCashInWalletAssetBalance('USDC', '-')
+    await savingsPage.expectSavingsBalance({ sDaiBalance: '9,332.66 sDAI', estimatedDaiValue: '10,000' })
+    await savingsPage.expectCashInWalletAssetBalance('DAI', '-')
   })
 })
