+++ packages/app/src/features/dialogs/savings/deposit/e2e/mainnet/DepositUSDT.test-e2e.ts
@@ -5,7 +5,7 @@ import { setup } from '@/test/e2e/setup'
 import { setupFork } from '@/test/e2e/setupFork'
 import { test } from '@playwright/test'
 import { mainnet } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../common/e2e/SavingsDialog.PageObject'
+import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
 
 test.describe('Deposit USDT on Mainnet', () => {
   const fork = setupFork({ blockNumber: 19990683n, chainId: mainnet.id })
