+++ packages/app/src/features/dialogs/savings/withdraw/e2e/gnosis/WithdrawXDAI.test-e2e.ts
@@ -5,7 +5,7 @@ import { setup } from '@/test/e2e/setup'
 import { setupFork } from '@/test/e2e/setupFork'
 import { test } from '@playwright/test'
 import { gnosis } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../common/e2e/SavingsDialog.PageObject'
+import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
 
 test.describe('Withdraw XDAI on Gnosis', () => {
   const fork = setupFork({
